+++ intro2ng6/src/app/app.component.ts
@@ -1,23 +1,14 @@
+import { enableProdMode } from '@angular/core';
 import { Component } from '@angular/core';
 
-function log(target, name, descriptor) {
-  // console.log(target, name, descriptor)
+import { environment } from '../environments/environment';
 
-  // store the original function in a var. 
-  const original = descriptor.value;
+if (environment.production) {
+  enableProdMode();
+}
 
-  // do some manipulations with the descriptor.value
-  descriptor.value = (...args) => {    
-    console.log("Arguments ", args, "were passed into this function.");
-    const result = original.apply(this, args); 
-    console.error("The results of this function is ", result);
-    return result;
-  }
-
-  // return the descriptor
-  // original();
-  console.log(target, name, descriptor);
-  return descriptor;
+function log(className) {
+  console.log(className)
 }
 
 @Component({
@@ -33,7 +24,6 @@ export class AppComponent {
     this.aSimpleMethod("This statement was generated by the constructor ", this.aSimpleMethod(5, 2));
   }
 
-  @log
   aSimpleMethod(a, b) {
     // console.log("what's the answer?");
     return a*a
